Backend/requirements.txt
asgiref==3.11.0
django==5.0
django-cors-headers==4.9.0
django-js-asset==3.1.2
django-prose-editor==0.23.1
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
faker==40.1.2
nh3==0.3.2
pillow==12.1.0
pyjwt==2.10.1
python-dotenv==1.2.1
sqlparse==0.5.5
psycopg2-binary==2.9.9
gunicorn==21.2.0
django-redis==5.4.0
dj-database-url==2.1.0
requests
